Don't access MWServices prematurely in Maintenence.php
In case of the referenced addWiki bug, the following was happening:
1. MWServices were initialized before Setup.php is complete
and before extensions were loaded. Deprecation warning was emitted.
2. onMWServices hook was run, thus hook container was created, but
without any extensions loaded.
3. When global MWServices instance was allowed, and services were
reset, hook container was salvaged. When that happens, already
instantiated handlers are preserved, so onMediaWikiServices handlers
(none at this point) were preserved.
4. GlobalPreferences refefine PreferencesFactory, but it was not
actually redefined, since onMediaWikiServices handler array was
salvaged in 3
5. The rest of the hooks work ok, since they are dynamically initialized
after extensions were loaded.
6. One of the hook handlers, implemented in GlobalPreferences,
expected PreferencesFactory to be redefined, but it was not.
7. Boom!

Ref T285894, T275453.

(cherry picked from commit e73769335d8e1f18fa64bcd56ce088c7a136723b)

+++ maintenance/includes/Maintenance.php
@@ -45,11 +45,12 @@ use Wikimedia\Rdbms\LBFactory;
  * bar is the option value of the option for param foo
  * baz is the arg value at index 0 in the arg list
  *
- * WARNING: the constructor, shouldExecute(), setup(), getName() and
- * loadSettings() are called before Setup.php is run, which means most of the
+ * WARNING: the constructor, shouldExecute(), setup(), finalSetup(), getName()
+ * and loadSettings() are called before Setup.php is complete, which means most of the
  * common infrastructure, like logging or autoloading, is not available. Be
  * careful when changing these methods or the ones called from them. Likewise,
- * be careful with the constructor when subclassing.
+ * be careful with the constructor when subclassing. MediaWikiServices instance
+ * is not yet available at this point.
  *
  * @stable for subclassing
  *
@@ -1194,10 +1195,15 @@ abstract class Maintenance {
 		}
 		if ( $this->hasOption( 'dbgroupdefault' ) ) {
 			$wgDBDefaultGroup = $this->getOption( 'dbgroupdefault', null );
-
-			$service = MediaWikiServices::getInstance()->peekService( 'DBLoadBalancerFactory' );
-			if ( $service ) {
-				$service->destroy();
+			// TODO: once MediaWikiServices::getInstance() starts throwing exceptions
+			// and not deprecation warnings for premature access to service container,
+			// we can remove this line. This method is called before Setup.php,
+			// so it would be guaranteed DBLoadBalancerFactory is not yet initialized.
+			if ( MediaWikiServices::hasInstance() ) {
+				$service = MediaWikiServices::getInstance()->peekService( 'DBLoadBalancerFactory' );
+				if ( $service ) {
+					$service->destroy();
+				}
 			}
 		}
 
@@ -1218,9 +1224,15 @@ abstract class Maintenance {
 				$wgLBFactoryConf['serverTemplate']['user'] = $wgDBuser;
 				$wgLBFactoryConf['serverTemplate']['password'] = $wgDBpassword;
 			}
-			$service = MediaWikiServices::getInstance()->peekService( 'DBLoadBalancerFactory' );
-			if ( $service ) {
-				$service->destroy();
+			// TODO: once MediaWikiServices::getInstance() starts throwing exceptions
+			// and not deprecation warnings for premature access to service container,
+			// we can remove this line. This method is called before Setup.php,
+			// so it would be guaranteed DBLoadBalancerFactory is not yet initialized.
+			if ( MediaWikiServices::hasInstance() ) {
+				$service = MediaWikiServices::getInstance()->peekService( 'DBLoadBalancerFactory' );
+				if ( $service ) {
+					$service->destroy();
+				}
 			}
 		}
